Update constructor types
