+++ models/user.js
@@ -1,14 +1,12 @@
-import Recipe from "./recipes";
-import Ingredient from "./ingredients";
-
 const mongoose = require(`mongoose`);
+const ObjectId = mongoose.Schema.Types.ObjectId;
 
 const userSchema = mongoose.Schema({
-    email: String,
-    password: String,
+    email: { String, required: true, unique: true },
+    password: { String, required: true },
     token: String,
-    recipes: Recipe,
-    ingredients: Ingredient,
+    recipes: ObjectId,
+    ingredients: ObjectId,
 });
 
 module.exports.User = mongoose.model(`User`, userSchema);
